create models path

+++ src/dataset.py
@@ -51,7 +51,7 @@ def load_h5(wild_path, num_points, k):
                     print('load', neighbours_file)
                     indices = np.load(neighbours_file)
                 else:
-                    indices = indices_k_neighbours(pcs, k)
+                    indices = indices_k_neighbours(pcs, k).astype(np.double)
                     np.save(neighbours_file, indices)
                     print("Indexes file created in: ", neighbours_file)
             pcs = np.dstack([pcs, indices])
